requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pydantic[email]>=2.6
email-validator>=2.1
python-multipart>=0.0.9
